mockmate-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.5
pydantic-settings==2.7.0
pydantic[email]==2.10.5
sqlalchemy==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
python-multipart==0.0.20
